+++ src/pages/_authenticated/_blank-layout/route.tsx
@@ -1,0 +1,13 @@
+import { Outlet, createFileRoute } from "@tanstack/react-router";
+
+export const Route = createFileRoute("/_authenticated/_blank-layout")({
+	component: BlankLayoutComponent,
+});
+
+function BlankLayoutComponent() {
+	return (
+		<div className="flex flex-col min-h-screen">
+			<Outlet />
+		</div>
+	)
+}
