+++ src/pages/_authenticated/route.tsx
@@ -1,0 +1,15 @@
+import { createFileRoute, Outlet, redirect } from "@tanstack/react-router";
+
+export const Route = createFileRoute("/_authenticated")({
+	beforeLoad: ({ context, location }) => {
+		if (!context.auth.isAuthenticated) {
+			throw redirect({
+				to: "/",
+				search: {
+					redirect: location.href,
+				},
+			});
+		}
+	},
+	component: () => <Outlet />,
+});
