--- src/pages/_fixed-header-layout/addfriend.tsx
@@ -1,9 +1,0 @@
-import { createFileRoute } from "@tanstack/react-router";
-
-export const Route = createFileRoute("/_fixed-header-layout/addfriend")({
-	component: RouteComponent,
-});
-
-function RouteComponent() {
-	return <div>Hello "/_fixed-header-layout/friend"!</div>;
-}
